\name{productsef}
\alias{productsef}                                                                                                  
\title{
this function is for internal use only
}
\description{
the function is a part of the function pkorner
}
\usage{
productsef(f, k, n, j)
}
\arguments{
  \item{f}{
  searcher efficiency
}
  \item{n}{
number of searches: n*d = duration of study
}
  \item{k}{
factor by which the search efficiency is multiplied after each search for a specific carcass (see details)
}
  \item{j}{
an index
}
}


\author{
Fraenzi Korner-Nievergelt
}


\seealso{
\code{\link{pkorner}}
}
\keyword{methods}
\keyword{misc}
