\name{carcass-package}
\alias{carcass-package}
\alias{carcass}
\docType{package}
\title{
Estimation of the number of fatalities from carcass searches
}
\description{
The number of bird or bat fatalities from collisions with buildings, towers or wind energy turbines can be estimated based on carcass searches and experimentally assessed carcass persistence times and searcher efficiency. This package provides diverse functions for estimating the probability that a bird or bat that died is found by a searcher. It further provides a function to obtain a posterior distribution of the number of fatalities based on the number of carcasses found and the estimated detection probability.
}
\details{
\tabular{ll}{
Package: \tab carcass\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-23\cr
License: \tab GUI2\cr
LazyLoad: \tab yes\cr
}
see worked example in Korner-Nievergelt et al 2011
}
\author{
Fraenzi Korner-Nievergelt, Ivo Niermann, Oliver Behr, Robert Brinkmann, Pius Korner, Barbara Hellriegel

Maintainer: Fraenzi Korner-Nievergelt <fraenzi.korner@oikostat.ch>

The authors thank Manuela Huso for commenting parts of the code.
}
\references{
Erickson WP, Jeffrey J, Kronner K, Bay K (2004) Stateline Wind Project Wildlife. Monitoring Final Report, July 2001 - December 2003. - Technical report peer-reviewed by and submitted to FPL Energy, the Oregon Energy Facility Siting Council, and the Stateline Technical Advisory Committee, Oregon, USA, 105 pp\cr
Huso M (2010) An estimator of wildlife fatality from observed carcasses. Environmetrics 22: 318-329\cr
Korner-Nievergelt F, Korner-Nievergelt P, Behr O, Niermann I, Brinkmann R, Hellriegel B (2011) A new method to determine bird and bat fatality at wind energy turbines from carcass searches. Wildlife Biology 17: 350-363
}
\keyword{ package }